refactor(conan): Introduce version-specific handler
This handler contain the calls and logic specific to a given Conan version.
The functions left in the Conan package manager are Conan version agnostic.
This is preliminary work for the support of Conan 2.
