+++ plugins/package-managers/conan/src/main/kotlin/Conan.kt
@@ -45,20 +45,15 @@ import org.ossreviewtoolkit.downloader.VcsHost
 import org.ossreviewtoolkit.downloader.VersionControlSystem
 import org.ossreviewtoolkit.model.Hash
 import org.ossreviewtoolkit.model.HashAlgorithm
-import org.ossreviewtoolkit.model.Identifier
-import org.ossreviewtoolkit.model.Package
-import org.ossreviewtoolkit.model.PackageReference
 import org.ossreviewtoolkit.model.Project
 import org.ossreviewtoolkit.model.ProjectAnalyzerResult
 import org.ossreviewtoolkit.model.RemoteArtifact
-import org.ossreviewtoolkit.model.Scope
 import org.ossreviewtoolkit.model.VcsInfo
 import org.ossreviewtoolkit.model.config.AnalyzerConfiguration
 import org.ossreviewtoolkit.model.config.Excludes
 import org.ossreviewtoolkit.plugins.api.OrtPlugin
 import org.ossreviewtoolkit.plugins.api.PluginDescriptor
 import org.ossreviewtoolkit.utils.common.CommandLineTool
-import org.ossreviewtoolkit.utils.common.Os
 import org.ossreviewtoolkit.utils.common.masked
 import org.ossreviewtoolkit.utils.common.safeDeleteRecursively
 import org.ossreviewtoolkit.utils.common.stashDirectories
@@ -108,19 +103,19 @@ class Conan(
     private val config: ConanConfig
 ) : PackageManager("Conan") {
     companion object {
-        private val DUMMY_COMPILER_SETTINGS = arrayOf(
+        internal val DUMMY_COMPILER_SETTINGS = arrayOf(
             "-s", "compiler=gcc",
             "-s", "compiler.libcxx=libstdc++",
             "-s", "compiler.version=11.1"
         )
 
-        private const val SCOPE_NAME_DEPENDENCIES = "requires"
-        private const val SCOPE_NAME_DEV_DEPENDENCIES = "build_requires"
+        internal const val SCOPE_NAME_DEPENDENCIES = "requires"
+        internal const val SCOPE_NAME_DEV_DEPENDENCIES = "build_requires"
     }
 
     override val globsForDefinitionFiles = listOf("conanfile*.txt", "conanfile*.py")
 
-    private val conanHome = Os.userHomeDirectory.resolve(".conan")
+    private val handler = ConanV1Handler(this)
 
     // This is where Conan caches downloaded packages [1]. Note that the package cache is not concurrent, and its
     // layout does not support packages from different remotes that are named (and versioned) the same.
@@ -129,7 +124,7 @@ class Conan(
     //
     // [1]: https://docs.conan.io/en/latest/reference/config_files/conan.conf.html#storage
     // [2]: https://docs.conan.io/en/latest/configuration/download_cache.html#download-cache
-    private val conanStoragePath = conanHome.resolve("data")
+    internal val conanStoragePath by lazy { handler.getConanStoragePath() }
 
     private val pkgInspectResults = mutableMapOf<String, JsonObject>()
 
@@ -171,7 +166,7 @@ class Conan(
         val conanConfig = sequenceOf(workingDir, analysisRoot).map { it.resolve("conan_config") }
             .find { it.isDirectory }
 
-        val directoryToStash = conanConfig?.let { conanHome } ?: conanStoragePath
+        val directoryToStash = conanConfig?.let { handler.getConanHome() } ?: conanStoragePath
 
         stashDirectories(directoryToStash).use {
             configureRemoteAuthentication(conanConfig)
@@ -181,43 +176,9 @@ class Conan(
                 config.lockfileName?.let { hasLockfile(workingDir.resolve(it).path) } == true
             }
 
-            val jsonFile = createOrtTempDir().resolve("info.json")
-            if (config.lockfileName != null) {
-                verifyLockfileBelongsToProject(workingDir, config.lockfileName)
-                ConanCommand.run(
-                    workingDir,
-                    "info", definitionFile.name,
-                    "-l", config.lockfileName,
-                    "--json", jsonFile.absolutePath
-                ).requireSuccess()
-            } else {
-                ConanCommand.run(
-                    workingDir,
-                    "info", definitionFile.name,
-                    "--json", jsonFile.absolutePath,
-                    *DUMMY_COMPILER_SETTINGS
-                ).requireSuccess()
-            }
+            val handlerResults = handler.process(definitionFile, config.lockfileName)
 
-            val pkgInfos = parsePackageInfos(jsonFile)
-            jsonFile.parentFile.safeDeleteRecursively()
-
-            val packageList = removeProjectPackage(pkgInfos, definitionFile.name)
-            val packages = parsePackages(packageList, workingDir)
-            val projectInfo = findProjectPackageInfo(pkgInfos, definitionFile.name)
-
-            val dependenciesScope = Scope(
-                name = SCOPE_NAME_DEPENDENCIES,
-                dependencies = parseDependencyTree(pkgInfos, projectInfo.requires, workingDir)
-            )
-            val devDependenciesScope = Scope(
-                name = SCOPE_NAME_DEV_DEPENDENCIES,
-                dependencies = parseDependencyTree(pkgInfos, projectInfo.buildRequires, workingDir)
-            )
-
-            val projectPackage = generateProjectPackage(projectInfo, definitionFile, workingDir)
-
-            return listOf(
+            val result = with(handlerResults) {
                 ProjectAnalyzerResult(
                     project = Project(
                         id = projectPackage.id,
@@ -235,11 +196,13 @@ class Conan(
                     ),
                     packages = packages.values.toSet()
                 )
-            )
+            }
+
+            return listOf(result)
         }
     }
 
-    private fun verifyLockfileBelongsToProject(workingDir: File, lockfileName: String?) {
+    internal fun verifyLockfileBelongsToProject(workingDir: File, lockfileName: String?) {
         require(workingDir.resolve(lockfileName.orEmpty()).canonicalFile.startsWith(workingDir.canonicalFile)) {
             "The provided lockfile path points to the directory outside of the analyzed project: '$lockfileName' and " +
                 "potentially does not belong to the project. Please move the lockfile to the '$workingDir' and " +
@@ -248,39 +211,14 @@ class Conan(
     }
 
     private fun configureRemoteAuthentication(conanConfig: File?) {
-        val remoteList = runCatching {
-            // Install configuration from a local directory if available.
-            conanConfig?.let {
-                ConanCommand.run("config", "install", it.absolutePath).requireSuccess()
-            }
-
-            // List configured remotes in "remotes.txt" format.
-            ConanCommand.run("remote", "list", "--raw").requireSuccess()
-        }.getOrElse {
-            logger.warn { "Failed to list remotes." }
-            return
+        // Install configuration from a local directory if available.
+        conanConfig?.let {
+            ConanCommand.run("config", "install", it.absolutePath).requireSuccess()
         }
 
-        val remotes = parseConanRemoteList(remoteList.stdout)
+        val remotes = handler.listRemotes()
         configureUserAuthentication(remotes)
     }
-
-    private fun parseConanRemoteList(remoteList: String): List<Pair<String, String>> =
-        remoteList.lines().mapNotNull { line ->
-            // Extract the remote URL.
-            val trimmedLine = line.trim()
-            if (trimmedLine.isEmpty() || trimmedLine.startsWith('#')) return@mapNotNull null
-
-            val wordIterator = trimmedLine.splitToSequence(' ').iterator()
-
-            if (!wordIterator.hasNext()) return@mapNotNull null
-            val remoteName = wordIterator.next()
-
-            if (!wordIterator.hasNext()) return@mapNotNull null
-            val remoteUrl = wordIterator.next()
-
-            remoteName to remoteUrl
-        }
 
     private fun configureUserAuthentication(remotes: List<Pair<String, String>>) =
         remotes.forEach { (remoteName, remoteUrl) ->
@@ -309,68 +247,15 @@ class Conan(
         }
 
     /**
-     * Return the dependency tree for the given [direct scope dependencies][requires].
-     */
-    private fun parseDependencyTree(
-        pkgInfos: List<PackageInfo>,
-        requires: List<String>,
-        workingDir: File
-    ): Set<PackageReference> =
-        buildSet {
-            requires.forEach { childRef ->
-                pkgInfos.find { it.reference.orEmpty() == childRef }?.let { pkgInfo ->
-                    logger.debug { "Found child '$childRef'." }
-
-                    val id = parsePackageId(pkgInfo, workingDir)
-                    val dependencies = parseDependencyTree(pkgInfos, pkgInfo.requires, workingDir) +
-                        parseDependencyTree(pkgInfos, pkgInfo.buildRequires, workingDir)
-
-                    add(PackageReference(id, dependencies = dependencies))
-                }
-            }
-        }
-
-    /**
-     * Return the map of packages and their identifiers which are contained in [pkgInfos].
-     */
-    private fun parsePackages(pkgInfos: List<PackageInfo>, workingDir: File): Map<String, Package> =
-        pkgInfos.associate { pkgInfo ->
-            val pkg = parsePackage(pkgInfo, workingDir)
-            "${pkg.id.name}:${pkg.id.version}" to pkg
-        }
-
-    /**
-     * Return the [Package] parsed from the given [pkgInfo].
-     */
-    private fun parsePackage(pkgInfo: PackageInfo, workingDir: File): Package {
-        val homepageUrl = pkgInfo.homepage.orEmpty()
-
-        val id = parsePackageId(pkgInfo, workingDir)
-        val conanData = readConanData(id.name, id.version, conanStoragePath)
-
-        return Package(
-            id = id,
-            authors = parseAuthors(pkgInfo),
-            declaredLicenses = pkgInfo.license.toSet(),
-            description = inspectField(pkgInfo.displayName, workingDir, "description").orEmpty(),
-            homepageUrl = homepageUrl,
-            binaryArtifact = RemoteArtifact.EMPTY, // TODO: implement me!
-            sourceArtifact = parseSourceArtifact(conanData),
-            vcs = processPackageVcs(VcsInfo.EMPTY, homepageUrl),
-            isModified = conanData.hasPatches
-        )
-    }
-
-    /**
      * Return the value `conan inspect` reports for the given [field], or null if the field does not exist.
      */
-    private fun inspectField(pkgName: String, workingDir: File, field: String): String? {
+    internal fun inspectField(pkgName: String, workingDir: File, field: String): String? {
         val results = pkgInspectResults.getOrPut(pkgName) {
             // Note: While Conan 2 supports inspect output to stdout, Conan 1 does not and a temporary file is required,
             // see https://github.com/conan-io/conan/issues/6972.
             val jsonFile = createOrtTempDir().resolve("inspect.json")
 
-            ConanCommand.run(workingDir, "inspect", pkgName, "--json", jsonFile.absolutePath).requireSuccess()
+            handler.runInspectCommand(workingDir, pkgName, jsonFile)
 
             Json.parseToJsonElement(jsonFile.readText()).jsonObject.also {
                 jsonFile.parentFile.safeDeleteRecursively()
@@ -383,35 +268,9 @@ class Conan(
     }
 
     /**
-     * Find the [PackageInfo] that represents the project defined in the definition file.
-     */
-    private fun findProjectPackageInfo(pkgInfos: List<PackageInfo>, definitionFileName: String): PackageInfo =
-        pkgInfos.first {
-            // Use "in" because conanfile.py's reference string often includes other data.
-            definitionFileName in it.reference.orEmpty()
-        }
-
-    /**
-     * Return the full list of packages, excluding the project level information.
-     */
-    private fun removeProjectPackage(pkgInfos: List<PackageInfo>, definitionFileName: String): List<PackageInfo> =
-        pkgInfos.minusElement(findProjectPackageInfo(pkgInfos, definitionFileName))
-
-    /**
-     * Return the [Identifier] for the package contained in [pkgInfo].
-     */
-    private fun parsePackageId(pkgInfo: PackageInfo, workingDir: File) =
-        Identifier(
-            type = "Conan",
-            namespace = "",
-            name = inspectField(pkgInfo.displayName, workingDir, "name").orEmpty(),
-            version = inspectField(pkgInfo.displayName, workingDir, "version").orEmpty()
-        )
-
-    /**
      * Return the [VcsInfo] contained in [pkgInfo].
      */
-    private fun parseVcsInfo(pkgInfo: PackageInfo): VcsInfo {
+    internal fun parseVcsInfo(pkgInfo: PackageInfo): VcsInfo {
         val revision = pkgInfo.revision.orEmpty()
         val url = pkgInfo.url.orEmpty()
         val vcsInfo = VcsHost.parseUrl(url)
@@ -422,7 +281,7 @@ class Conan(
      * Return the source artifact contained in [conanData], or [RemoteArtifact.EMPTY] if no source artifact is
      * available.
      */
-    private fun parseSourceArtifact(conanData: ConanData): RemoteArtifact {
+    internal fun parseSourceArtifact(conanData: ConanData): RemoteArtifact {
         val url = conanData.url ?: return RemoteArtifact.EMPTY
         val hashValue = conanData.sha256.orEmpty()
         val hash = Hash.NONE.takeIf { hashValue.isEmpty() } ?: Hash(hashValue, HashAlgorithm.SHA256)
@@ -431,65 +290,35 @@ class Conan(
     }
 
     /**
-     * Return a [Package] containing project-level information from [pkgInfo] and [definitionFile] using the
-     * `conan inspect` command if possible:
-     * - conanfile.txt: `conan inspect conanfile.txt` is not supported.
-     * - conanfile.py: `conan inspect conanfile.py` is supported and more useful project metadata is parsed.
-     *
-     * TODO: The format of `conan info` output for a conanfile.txt file may be such that we can get project metadata
-     *       from the `requires` field. Need to investigate whether this is a sure thing before implementing.
-     */
-    private fun generateProjectPackage(pkgInfo: PackageInfo, definitionFile: File, workingDir: File): Package {
-        fun inspectPyFile(field: String) =
-            definitionFile.name.takeIf { it == "conanfile.py" }?.let { inspectField(it, workingDir, field) }
-
-        return Package(
-            id = Identifier(
-                type = projectType,
-                namespace = "",
-                name = inspectPyFile("name") ?: pkgInfo.reference.orEmpty(),
-                version = inspectPyFile("version").orEmpty()
-            ),
-            authors = parseAuthors(pkgInfo),
-            declaredLicenses = pkgInfo.license.toSet(),
-            description = inspectPyFile("description").orEmpty(),
-            homepageUrl = pkgInfo.homepage.orEmpty(),
-            binaryArtifact = RemoteArtifact.EMPTY, // TODO: implement me!
-            sourceArtifact = RemoteArtifact.EMPTY, // TODO: implement me!
-            vcs = parseVcsInfo(pkgInfo)
-        )
-    }
-
-    /**
      * Parse information about the package author from the given [package info][pkgInfo]. If present, return a set
      * containing the author name; otherwise, return an empty set.
      */
-    private fun parseAuthors(pkgInfo: PackageInfo): Set<String> =
+    internal fun parseAuthors(pkgInfo: PackageInfo): Set<String> =
         parseAuthorString(pkgInfo.author).mapNotNullTo(mutableSetOf()) { it.name }
+
+    internal fun readConanData(name: String, version: String, conanStorageDir: File): ConanData {
+        val conanDataFile = handler.getConanDataFile(name, version, conanStorageDir)
+        val root = Yaml.default.parseToYamlNode(conanDataFile.readText()).yamlMap
+
+        val patchesForVersion = root.get<YamlMap>("patches")?.get<YamlList>(version)
+        val hasPatches = !patchesForVersion?.items.isNullOrEmpty()
+
+        val sourceForVersion = root.get<YamlMap>("sources")?.get<YamlMap>(version)
+        val sha256 = sourceForVersion?.get<YamlScalar>("sha256")?.content
+
+        val url = sourceForVersion?.get<YamlNode>("url")?.let {
+            when {
+                it is YamlList -> it.yamlList.items.firstOrNull()?.yamlScalar?.content
+                else -> it.yamlScalar.content
+            }
+        }
+
+        return ConanData(url, sha256, hasPatches)
+    }
 }
 
-private data class ConanData(
+internal data class ConanData(
     val url: String?,
     val sha256: String?,
     val hasPatches: Boolean
 )
-
-private fun readConanData(name: String, version: String, conanStorageDir: File): ConanData {
-    val conanDataFile = conanStorageDir.resolve("$name/$version/_/_/export/conandata.yml")
-    val root = Yaml.default.parseToYamlNode(conanDataFile.readText()).yamlMap
-
-    val patchesForVersion = root.get<YamlMap>("patches")?.get<YamlList>(version)
-    val hasPatches = !patchesForVersion?.items.isNullOrEmpty()
-
-    val sourceForVersion = root.get<YamlMap>("sources")?.get<YamlMap>(version)
-    val sha256 = sourceForVersion?.get<YamlScalar>("sha256")?.content
-
-    val url = sourceForVersion?.get<YamlNode>("url")?.let {
-        when {
-            it is YamlList -> it.yamlList.items.firstOrNull()?.yamlScalar?.content
-            else -> it.yamlScalar.content
-        }
-    }
-
-    return ConanData(url, sha256, hasPatches)
-}
